requirements/prod.txt
mangum==0.10.0
fastapi==0.65.0
uvicorn==0.13.3
psycopg2-binary==2.8.6